Day 1 - Experimenting with p5.play
Today was about setting up the template for the project, today was mostly troubleshooting and downloading the p5.play library. Then testing it out and learning the different functions and methods inside the external library while trying to implement it all into a class. My goal is to implement physics and have the objects move and interact with the canvas and possible other objects for tomorrow. Challenge is figuring out how and what methods to use for this library, as it is quite a lot different than what we have learned in class. But so far, I would say this library is a great tool for my project.