Day 3  - Creating the template for the game and furhter experimenting with p5
Today was a pretty good day. Got a border set up for the game, got the basic game physics working with balls. Now goal is to change it to images, which i did near the end. Using minecraft diamonds to experiment with the images. It's rather difficult as the collider does not have the same dimensions as the image, so need to figure that out.