Day 5 - A Bit of a set back
Today was a set back as I figured that it would be better to implement classes into this assingment, however its quite difficult learning how to use the sprites with classes so today had to rewrite almost all of my code, but i think this should be helpful in the future. I also decided to make it easier for myself and figure out how to combine just the balls themselves, before i figure out how to combine images.