Day 6 - Implemented groups, and pushed the balls into an array -
Today was pretty good learned a lot about the p5.play library. The only problem if find is implementing this library with other java script functions. Especially the class method, it's very difficult, thus I'm still deciding weather or not I should continue with the class method, or rather just use the sole concepts of the p5.play library. Thus, I have decided to do some testing on a different folder and have implemented a merge function that when the balls collide they merge together. This was a lot easier to complete without the class, so im  wondering if it would be more beneficial to scrap the class idea entirely.