Day 10 - Was a little bit of progress not as much as i hope for
Today tried fixing the restart function wasn't really able to figure it out for today, so implemented a score (still need to figure out how to draw it above the sprites) as right now it is beign drawn behind the sprites. Currently the code is all over the place i need to be little more organized and work on one thing at a time as right now kinda difficult to figure everything out all at once.