Day 12  - added more images and got the coliders (somewhat) working with the images
Today I just added more images to the library, found a debug function that made figuring out the collider scale for the images so much easier. Got it working pretty well, only thing is once the fruit images get bigger the collider doesn't and makes it look like the fruits are floating. Then also need to figure out the image index properly as sometimes the fruits repeat when merged rather than going to the next fruit.